fix: remove unnecessary redisConfig
